Add internal annotation in core_impl, move @Internal in common.
