+++ core_impl/build.gradle
@@ -10,3 +10,5 @@ dependencies {
 
     signature "org.codehaus.mojo.signature:java16:+@signature"
 }
+
+javadoc.exclude 'com/google/instrumentation/internal/**'
